ECOSCOM-4134 - Alf roles priority for case feature

+++ idocs-repo/src/main/java/ru/citeck/ecos/icase/CaseStatusServiceImpl.java
@@ -67,7 +67,9 @@ public class CaseStatusServiceImpl implements CaseStatusService {
     private void clearBeforeCaseStatus(NodeRef caseRef) {
         NodeRef beforeCaseStatus = caseStatusAssocDao.getStatusByAssoc(caseRef, ICaseModel.ASSOC_CASE_STATUS_BEFORE);
         if (isAlfRef(beforeCaseStatus)) {
-            nodeService.removeAssociation(caseRef, beforeCaseStatus, ICaseModel.ASSOC_CASE_STATUS_BEFORE);
+            if (beforeCaseStatus != null) {
+                nodeService.removeAssociation(caseRef, beforeCaseStatus, ICaseModel.ASSOC_CASE_STATUS_BEFORE);
+            }
         } else {
             nodeService.setProperty(caseRef, ICaseModel.ASSOC_CASE_STATUS_BEFORE_PROP, null);
         }
